Front end

See merge request s07-webmobile2-sub2/S07P12A201!218

+++ front-end/queant/src/store/modules/portfolio.js
@@ -18,8 +18,6 @@ export default {
     customProducts: state => state.customProducts,
     portfolios: state => state.portfolios,
     comparisonPortfolio: state => state.comparisonPortfolio,
-    newlyAddedPortfolio: state => state.newlyAddedPortfolio,
-    deletedPortfolio: state => state.deletedPortfolio,
   },
   mutations: {
     SET_CUSTOM_PRODUCTS: (state, customProducts) => state.customProducts = customProducts,
@@ -27,21 +25,18 @@ export default {
     // Comparison Portfolio
 
     UPDATE_CPORTFOLIO_FROM_DB(state, value) {state.comparisonPortfolio = value;},
-    
-      // DB정리
-    CLEAR_CPORTFOLIO_DB(state) {
-      state.newlyAddedPortfolio = [],
-      state.deletedPortfolio = []
-    },
+
     ADD_COMPARISON_PORTFOLIO_PRODUCTS(state, product) {
       state.comparisonProducts.push(product);
     },
       // 포트폴리오 삭제하기
     POP_CPORTFOLIO(state, cportfolio_cnt) {
-      state.deletedPortfolio.push(cportfolio_cnt)
       let portfolioNo = cportfolio_cnt
       // 삭제된 포트폴리오의 이후 포트폴리오의 cnt를 1씩 차감
       for (let i=portfolioNo; i < state.comparisonPortfolio.length; i++) {
+        for (let j=0; j<state.comparisonPortfolio[i].products.length; j++) {
+          state.comparisonPortfolio[i].products[j].portfolio_no -= 1
+        }
         state.comparisonPortfolio[i].cportfolio_cnt -= 1
         console.log(state.comparisonPortfolio[i])
       }
@@ -52,7 +47,6 @@ export default {
     ADD_COMPARISON_PORTFOLIO(state) {
       // 0개일 때
       if (state.comparisonPortfolio.length === 0) {
-        state.newlyAddedPortfolio.push(1);
         state.comparisonPortfolio.push({
           cportfolio_cnt: 1,
           products: []
@@ -60,7 +54,6 @@ export default {
         console.log('새로운 가상 포트폴리오가 추가되었습니다.')
       // 1~3개일 때
       } else if (state.comparisonPortfolio.length < 4) {
-        state.newlyAddedPortfolio.push(state.comparisonPortfolio.length+1)
         state.comparisonPortfolio.push({
           cportfolio_cnt: (state.comparisonPortfolio.length)+1,
           products: []
@@ -97,7 +90,7 @@ export default {
         alert(`${product.product.name}은 이미 포트폴리오에 있는 상품입니다.`)
       } else {
         state.comparisonPortfolio[portfolioNo-1].products.push(pushproduct)
-        alert(`${portfolioNo}번 포트폴리오에 ${pushproduct.name}상품이 담겼습니다.`)
+        alert(`${portfolioNo}번 포트폴리오에 ${product.product.name}상품이 담겼습니다.`)
       }
     },
     // 상품 삭제
@@ -105,9 +98,9 @@ export default {
     let portfolioNo = value[0]
     let FindProduct = value[1]
     let cportfolios = state.comparisonPortfolio
-    let item = cportfolios[portfolioNo-1].products.indexOf('FindProduct')
-    
-    alert(`${portfolioNo}번 포트폴리오의 ${FindProduct.name} 상품을 삭제했습니다.`)
+    let item = cportfolios[portfolioNo-1].products.indexOf(FindProduct)
+    console.log(item)
+    alert(`${portfolioNo}번 포트폴리오의 ${FindProduct.product.name} 상품을 삭제했습니다.`)
     cportfolios[portfolioNo-1].products.splice(item, 1)
     },
 
@@ -150,21 +143,24 @@ export default {
   actions: {
     // Custom Product
     deleteCustomProduct({ dispatch }, product_id) {
-      console.log(product_id)
-      axios({
-        url: spring.portfolio.custom(),
-        method: 'delete',
-        data: {
-          product_id: product_id
-        }
-      })
-      .then(res => {
-        console.log(res)
-        dispatch('fetchMyPortfolio')
-      })
-      .catch(err => {
-        console.log(err)
-      })
+      if (confirm('해당 상품을 삭제하시겠습니까?') === true) {
+        axios({
+          url: spring.portfolio.custom(),
+          method: 'delete',
+          data: {
+            product_id: product_id
+          }
+        })
+        .then(res => {
+          console.log(res)
+          dispatch('fetchMyPortfolio')
+        })
+        .catch(err => {
+          console.log(err)
+        })
+      } else {
+        return
+      }
     },
     updateCustomProduct({ dispatch }, payload) {
       payload.fixed_rsrv = false
@@ -325,95 +321,16 @@ export default {
     // Comparison Portfolio
 
     // 가상 포트폴리오 DB 저장 및 가져오기
-    saveToDb({ dispatch, commit, state, getters }) {
-      //삭제된 포트폴리오 DB에서 제거하기
-      // 삭제된 포트폴리오가 있다면 요청 보내기
-      if (state.deletedPortfolio.length != 0) {
-        for (let i=0; i<state.deletedPortfolio.length; i++) {
-          let port_no = state.deletedPortfolio[i]
-          // port_no번에 해당하는 포트폴리오를 모두 삭제해 주세요.
-          axios({
-            url: spring.portfolio.portfolio(),
-            method: 'delete',
-            data: {
-              member_id: getters.userInfo.member_id,
-              portfolio_no:port_no
-            }
-          }).then(res => {
-            console.log(res)
-            console.log(`${i+1}번 포트폴리오가 DB에서 삭제되었습니다.`)
-          })
-          .catch(err => {
-            console.log(err)
-          })
-        }
-      }
-    //현재 남아있는 포트폴리오 수정 및 삽입
-    for (let i = 0; i < state.comparisonPortfolio.length; i++){
-      // 새로 추가된 포트폴리오가 있다면
-      if (state.newlyAddedPortfolio.includes(i+1)) {
-          // 만약 비어있는 포트폴리오가 있다면 새로 저장할 수 없음.
-          if (state.comparisonPortfolio[i].cportfolio_cnt === (i+1) && state.comparisonPortfolio[i].products.length === 0) {
-            commit('POP_CPORTFOLIO', i+1)
-            alert(`비어 있는 ${i+1}번 포트폴리오를 삭제했습니다.`)
-          } else if (state.comparisonPortfolio[i].cportfolio_cnt === (i+1) && state.comparisonPortfolio[i].length != 0) {
-          let tempPortfolioList = [];
-          for (let port of state.comparisonPortfolio[i].products) {
-            let temp = {
-              amount: port.amount,
-              condition_ids: port.condition_ids,
-              end_date: port.end_date,
-              portfolio_no: port.portfolio_no,
-              product_id: port.product_id,
-              option_id: port.option_id,
-              start_date: port.start_date,
-            }
-            tempPortfolioList.push(temp)
-          }
-          console.log(tempPortfolioList)
-          console.log('올릴 자료')
-          axios({
-            url: spring.portfolio.portfolio(),
-            method: 'post',
-            data: {
-              member_id: getters.userInfo.member_id,
-              portfolio_no: i+1,
-              portfolio_dto_list: tempPortfolioList
-            }
-          }).then(res => {
-            console.log(res)
-            alert(`${i+1}번 포트폴리오가 DB에 새롭게 저장되었습니다.`)
-          })
-          .catch(err => {
-            console.log(err)
-          })}
-        // 새로 추가된 것이 없는 경우
-      } else {
-        // 이 경우에도 비어있는 경우 저장할 수 없고, 오히려 삭제할 것.
-        if (state.comparisonPortfolio[i].cportfolio_cnt === (i+1) && state.comparisonPortfolio[i].products.length === 0) {
-          alert(`비어 있는 포트폴리오를 저장할 수 없습니다. ${i+1}번 포트폴리오를 DB에서 삭제합니다.`)
-          let port_no = state.comparisonPortfolio[i].cportfolio_cnt
-          // port_no번에 해당하는 포트폴리오를 모두 삭제해 주세요.
-          axios({
-            url: spring.portfolio.portfolio(),
-            method: 'delete',
-            data: {
-              member_id: getters.userInfo.member_id,
-              portfolio_no:port_no
-            }
-          }).then(res => {
-            console.log(res)
-            alert(`${i+1}번 포트폴리오가 DB에서 삭제되었습니다.`)
-          })
-          .catch(err => {
-            console.log(err)
-          })
-          // 비어있지 않은 경우
-        } else if (state.comparisonPortfolio[i].cportfolio_cnt === (i+1) && state.comparisonPortfolio[i].length != 0) {
-        let tempPortfolioList = [];
-        console.log(state.comparisonPortfolio[i])
-        for (let port of state.comparisonPortfolio[i].products) {
-          console.log(port.option_id)
+    saveToDb({ commit, getters }) {
+      let tempPortfolioList = []
+      let checked_value = 0
+      for (let i = 0; i < getters.comparisonPortfolio.length; i++) {
+        if (getters.comparisonPortfolio[i].products.length === 0) {
+          commit('POP_CPORTFOLIO', i+1)
+          checked_value += 1
+          break
+        } else {
+        for (let port of getters.comparisonPortfolio[i].products) {
           let temp = {
             amount: port.amount,
             condition_ids: port.condition_ids,
@@ -424,29 +341,29 @@ export default {
             start_date: port.start_date,
           }
           tempPortfolioList.push(temp)
+        }}}
+      if (checked_value >= 1) {
+        alert('빈 포트폴리오가 있습니다. 저장이 되지 않았습니다.')
+        return
+      } else {
+      console.log(tempPortfolioList)
+      axios({
+        url: spring.portfolio.portfolio(),
+        method: 'post',
+        data: {
+          member_id: getters.userInfo.member_id,
+          portfolio_dto_list: tempPortfolioList
         }
-        console.log(tempPortfolioList)
-        axios({
-          url: spring.portfolio.portfolio(),
-          method: 'put',
-          data: {
-            member_id: getters.userInfo.member_id,
-            portfolio_no: i+1,
-            portfolio_dto_list: tempPortfolioList
-          }
-        }).then(res => {
-          console.log(res)
-          alert(`${i+1}번 포트폴리오를 DB에서 수정했습니다.`)
-        })
-        .catch(err => {
-          console.log(err)
-        })}
-      }
-    }
+      }).then(res => {
+        console.log(res)
+        alert('포트폴리오가 DB에 새롭게 저장되었습니다.')
+      })
+      .catch(err => {
+        console.log('포트폴리오 전송 실패')
+        console.log(err)
+      })}
     },
     getFromDb({ commit, dispatch, state, getters, actions }) {
-      // Clear DB를 실행(추가, 삭제 초기화)
-      commit('CLEAR_CPORTFOLIO_DB');
       // 포트폴리오 초기화
       commit('CLEAR_CPORTFOLIOS');
       // 정보 받아오기
@@ -491,7 +408,6 @@ export default {
               term: firstdata[i][j].option.save_term,
               rate_type: firstdata[i][j].option.rate_type
             })
-            dispatch('getProductDetails', firstdata[i][j].product_id)
           }
         }
         // ComparisonPortfolio를 업데이트
@@ -501,21 +417,5 @@ export default {
         console.log(err)
       })
     },
-    getProductDetails({ commit, dispatch }, productId) {
-      axios({
-        url: spring.product.detail(productId),
-        method: 'get',
-        params: {
-            productId: productId
-        }
-      })
-      .then(res=> {
-        console.log(res.data)
-        commit('ADD_COMPARISON_PORTFOLIO_PRODUCTS', res.data)
-      })
-      .catch(err => {
-        console.log(err)
-      })
-    }
   }
 };
